Videos/requirements.txt
openai>=1.52.0
python-dotenv>=1.0.1
requests>=2.31.0
pillow>=10.0.0
